requirements.txt
autopep8==1.2.1
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
inflect==0.2.5
marshmallow==2.3.0
marshmallow-jsonapi==0.3.0
PyJWT==1.4.0
PyMySQL==0.6.6
PyYAML==3.11
psycopg2==2.6
pycrypto==2.6.1
uWSGI==2.0.11.2

